% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Create bar plots}
\usage{
bar_plot(
  df,
  y = "Time",
  time_range = NULL,
  group_col = "Date",
  block_col = "Block",
  treat_col = "Treatment",
  treatment_level = NULL,
  y_lab = NULL,
  plot_title = "My plot",
  save = "FALSE",
  plot_width = 1000,
  plot_height = 651,
  palette = colorblind_palettes("wong"),
  ...
)
}
\arguments{
\item{df}{data frame or tibble that contains the data. Ideally from our lab template}

\item{y}{character (name) or number (index) for the column containing the variable of interest}

\item{time_range}{can assume values NULL (default), when not filtering per date. or
take one single value "2020-12-31", or an array of two values indicating the start and end
c("2020-12-31", "2021-01-31"). The extremes of the range are included. Use only when not the group_col
is a date column}

\item{group_col}{character (name) or number (index) for the column containing the faceting
variable used to split the plots in different panels}

\item{block_col}{character (name) or number (index) for the column containing the Blocking variable. Default "Block".}

\item{treat_col}{character (name) or number (index) for the column containing the Treatment variable. Default "Treatment".}

\item{treatment_level}{\code{NULL} (default) or an array giving a defined order for the treatment level that would make sense for plotting}

\item{y_lab}{string or \code{NULL} for name of y axis in the plot}

\item{plot_title}{string}

\item{save}{\code{TRUE} or \code{FALSE} to save the file to a png}

\item{plot_width}{the width of the plot, default to px unless units is specified see \code{png()}}

\item{plot_height}{the height of the plot, default to px unless units is specified see \code{png()}}

\item{...}{additional argument to be passed to the \link[grDevices]{png} function for saving the plot}
}
\value{
a ggplot object
}
\description{
\code{trend_plot} creates standard bar plots with standard error of the mean bars.
Multiple panels (facets) can be produced with the \code{group_col} argument
}
