% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_group_anovas.R
\name{by_group_anovas}
\alias{by_group_anovas}
\title{Perform anovas and multiple comparison tests on subset of a data frame}
\usage{
by_group_anovas(
  formula = SWP ~ Treatment + Block,
  df,
  time_range = NULL,
  group_col = "Date",
  treat_col = "Treatment",
  treatment_level = NULL,
  what = "summary",
  return_models = FALSE
)
}
\arguments{
\item{formula}{formula to be passed to the \code{stats::aov} function}

\item{df}{data frame. Ideally from the lab template}

\item{time_range}{NULL(default) do not select specific time range. Or take one single value "2020-12-31" for selecting a specific date,
or a vector of two values indicating the start and end c("2020-12-31", "2021-01-31"). The extremes of the range are included. Use only when group_col is a date column.}

\item{group_col}{column specifying the grouping variable. E.g. the name (character) or index (number)
of the date column for analysis repeated over time.}

\item{treat_col}{treatment column as name (character) or index (number)}

\item{treatment_level}{vector of characters corresponding to the desired order of levels
in \code{treat_col}. The control should be the first level}

\item{what}{can have values "summary" for anova summaries only,
"tukey" for Tukey HSD, "dunnett" for Dunnett comparisons against the one treatment (the ctrl)
and "all" for all of the above. Print to the console.}

\item{return_models}{TRUE or FALSE (default) to return a list of the anova models for further analysis}
}
\value{
a list of anova models when \code{return_models} is true
}
\description{
This function performs anova statistics for a full dataset split by user defined groups (e.g. Dates)
for trend measurements. It also provides two kinds of multiple comparison tests
Dunnet against the control and Tukey HSD.
}
\examples{

#for RCBD run (here an example from infiltration measurements in the AVF Salt Stress project)
by_group_anovas(formula = Time ~ Treatment + Block, df,
date_col = "Date", treat_col = "Treatment",
treatment_level = c("CTRL", "50G", "100G", "200G", "100GC", "100A"),
what = "all", return_models = FALSE)
}
