% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{Descriptive statistics of the data set by Date, Treatment, Block}
\usage{
data_check(
  df,
  y = "SWP",
  time_range = NULL,
  what = "Date-Treatment",
  date_col = "Date",
  block_col = "Block",
  treat_col = "Treatment",
  write_file = FALSE,
  file_name = "output.xlsx"
)
}
\arguments{
\item{df}{data frame or tibble that contains the data. Ideally from our lab template}

\item{y}{character (name) or number (index) for the column containing the variable of interest}

\item{time_range}{can assume values NULL (default), when not filtering per date, or
take one single value "2020-12-31", or an array of two values indicating the start and end
c("2020-12-31", "2021-01-31"). The extremes of the range are included.}

\item{what}{a string with values "Date", "Date-Treatment", "Date-Treatment-Block"}

\item{date_col}{character (name) or number (index) for the column containing the Date variable. Default "Date".}

\item{block_col}{character (name) or number (index) for the column containing the Blocking variable. Default "Block".}

\item{treat_col}{character (name) or number (index) for the column containing the Treatment variable. Default "Treatment".}
}
\value{

}
\description{
This is a function to check the quality of the data during the acquisition season
}
