% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_library_for_plotting.R
\name{exp_lib}
\alias{exp_lib}
\title{Expression library for plotting}
\usage{
exp_lib(what)
}
\arguments{
\item{what}{specify the expression you want. Use list_all to print the options to console}
}
\description{
This is a list of common measurement units with special characters that are useful
when passing labels to plots.
}
