% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swp.R
\name{read_swp}
\alias{read_swp}
\title{Read stem water potential data from lab template}
\usage{
read_swp(path, ...)
}
\arguments{
\item{path}{a string giving the full path to the file, make sure to replace "\" with "/" if in windows.}

\item{...}{additional arguments to be given to \code{read_xlsx()}}
}
\value{
a tibble
}
\description{
Read stem water potential data from lab template
}
