% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error of the Mean}
\usage{
se(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{TRUE or FALSE, default to TRUE}
}
\description{
Calculate the standard error of the mean. Avoid the need for calling a package just for it.
}
