% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swp_trend.R
\name{swp_trend}
\alias{swp_trend}
\title{Plot stem water potential data}
\usage{
swp_trend(df, y = "SWP", ...)
}
\arguments{
\item{df}{a tibble or data frame as the one coming from the function read_swp}

\item{y}{character (name) or number (index) for the column containing the variable of interest. Default "SWP", the name of the column in the lab template.}

\item{...}{additional argumnets to be passed to \link[brillab]{trend_plot}()}
}
\value{
a ggplot object
}
\description{
A wrapper for trend_plot for the common case of plotting stem water potentials
}
