% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_plot.R
\name{trend_plot}
\alias{trend_plot}
\title{Create trend plots}
\usage{
trend_plot(
  df,
  y = "SWP",
  date_col = "Date",
  block_col = "Block",
  treat_col = "Treatment",
  time_range = "all",
  treatment_level = NULL,
  y_lab = "SWP",
  plot_title = "My plot",
  save = "FALSE",
  plot_width = 1000,
  plot_height = 651,
  palette = colorblind_palettes("wong"),
  ...
)
}
\arguments{
\item{df}{data frame or tibble that contains the data. Ideally from our lab template}

\item{y}{character (name) or number (index) for the column containing the variable of interest}

\item{date_col}{character (name) or number (index) for the column containing the Date variable. Default "Date".}

\item{block_col}{character (name) or number (index) for the column containing the Blocking variable. Default "Block".}

\item{treat_col}{character (name) or number (index) for the column containing the Treatment variable. Default "Treatment".}

\item{time_range}{can assume values \code{"all"} (default), when not filtering per date, or take on single
value indicating the starting date \code{"2020-12-31"}, or an array of two values indicating the start and end
\code{c("2020-12-31", "2021-01-31")}. The extremes of the range are included.}

\item{treatment_level}{\code{NULL} (default) or an array giving a defined order for the treatment level that would make sense for plotting}

\item{y_lab}{string or \code{NULL} for name of y axis in the plot}

\item{plot_title}{string}

\item{save}{\code{TRUE} or \code{FALSE} to save the file to a png}

\item{plot_width}{the width of the plot, default to px unless units is specified see \code{png()}}

\item{plot_height}{the height of the plot, default to px unless units is specified see \code{png()}}

\item{...}{additional argument to be passed to the \link[grDevices]{png} function for saving the plot}
}
\description{
\code{trend_plot} creates standard line plots with transparent ribbons for showing standard
error of the mean to use with temporal data. Also check helpers such as \code{swp_trend}
for most common uses.
}
\details{
TO DO: currently this function does require a block column (although this is not needed for the plot)
If needed include support for when block column is missing.
TO DO: use the y to name the axis unless y_lab is assigned
}
